requirements.txt
flask==2.3.3
flask-cors==4.0.0
speechrecognition==3.10.0
pyttsx3==2.90
pillow==10.0.0
pytesseract==0.3.10
opencv-python==4.8.1.78
pyaudio==0.2.11
requests==2.31.0
python-docx==0.8.11
PyPDF2==3.0.1
numpy==1.24.3
openai==1.3.0
transformers==4.35.0
torch==2.1.0
websockets==11.0.3
asyncio==3.4.3
threading==1.0.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
